restructure api folders
