+++ .coveralls.yml
@@ -1,0 +1,1 @@
+repo_token: iYbUQ1Vu2OsfGAXT7iN2yYcJpTVjZa2KN
