+++ package.json
@@ -29,13 +29,13 @@
     "node": "8.11.3"
   },
   "scripts": {
-    "test": "npm run build && mocha ./build/test/app.spec.js  --exit || true",
+    "test": "nyc mocha ./build/test/app.test.js --exit",
     "lint": "eslint ./api",
     "build": "babel --presets es2015 ./api -d ./build/",
-    "start": "npm run build && node ./build/app.js",
-    "devStart": "npm run build && nodemon ./build/app.js",
-    "coverage": "istanbul cover ./build/test/app.spec.js",
-    "coveralls": "cat ./coverage/lcov.info | coveralls"
+    "start": "npm run build && node ./build/src/app.js",
+    "devStart": "npm run build && nodemon ./build/src/app.js",
+    "coverage": "nyc report ./build/test/app.test.js",
+    "coveralls": "--reporter=text-lcov | coveralls"
   },
   "repository": {
     "type": "git",
